requirements.txt
langchain
chromadb==0.3.26
urllib3
PyMuPDF
python-dotenv
unstructured
extract-msg
tabulate
pandoc
pypandoc
tqdm
werkzeug
flask
pymongo
openai
tiktoken
uuid
flask_jwt_extended
flask_cors
markdown
pillow